Challenge 2 complete.
Isolated crew member scroll view into its own file. Removed unnecessary code from mission view.
